fix(store): cast series_year to bigint in the postgres series-match query
series_needing_match aggregates media_items.series_year, which is INT4 on
postgres, but the query decoded it as Option<i64> (INT8). sqlx-postgres is
strict about column types, so every series-enrichment pass aborted in prod with
"Rust type Option<i64> (as SQL type INT8) is not compatible with SQL type INT4"
— shows never got enriched (movies/episodes were unaffected). sqlite is loose
and accepted it, so the in-tree sqlite tests never caught it.

Cast MAX(series_year)::bigint in the postgres query so the decode holds
regardless of the source column's width. Add a SeriesMetadataStore section to
the shared backend_conformance walk (runs against BOTH engines when
PHAROS_TEST_POSTGRES_URL is set) exercising series_needing_match end-to-end, so
a type divergence like this fails CI on postgres instead of only in prod.

+++ crates/pharos-store-sqlx/src/postgres.rs
@@ -3047,11 +3047,15 @@ impl SeriesMetadataStore for PostgresStore {
         // Distinct shows from their episodes' SeriesInfo, LEFT-JOINed against
         // series_metadata so never-matched / TTL-expired rows qualify and
         // `manual` overrides are excluded. Mirrors the sqlite query.
+        //
+        // media_items.series_year is INT4; cast the aggregate to bigint so the
+        // Option<i64> decode holds (sqlx-postgres is strict — INT4≠INT8 — and
+        // sqlite, being loose, accepts either).
         let rows: Vec<(String, String, Option<i64>)> = sqlx::query_as(
             "SELECT s.series_key, s.series_name, s.series_year \
              FROM ( \
                SELECT COALESCE(series_folder, series_name) AS series_key, \
-                      MIN(series_name) AS series_name, MAX(series_year) AS series_year \
+                      MIN(series_name) AS series_name, MAX(series_year)::bigint AS series_year \
                FROM media_items \
                WHERE kind = 'episode' AND series_name IS NOT NULL AND series_name <> '' \
                GROUP BY COALESCE(series_folder, series_name) \
